shard: consolidate db into ShardBlockChain
